[ui] Sensor history: Fix tick ID variable (#23802)
## Summary & Motivation

There is a bug on the sensor tick table where clicking to view a run
requested for a tick can show information about an unrelated tick.

This is because sensor tick IDs can be larger than JavaScript's
`Number.MAX_SAFE_INTEGER`, but they're being cast to `Number` for the
GraphQL variable. JS may cast the value to a number that doesn't match
the intended value, resulting in the wrong tick ID being sent in the
query.

We shouldn't be trying to represent them as `number` in JS. The GraphQL
input type expects it to be a BigInt, and we can just pass a string for
that.

## How I Tested These Changes

View sensor tick history with ticks that have tick IDs being cast to
incorrect values. Verify that the values are now correct, and that when
I click to view runs requested for the tick, the query variable and
rendered output are correct.