[dagster-sdf] Enable Caching & Asset Selection (#23750)
## Summary & Motivation

This PR enables caching simply by upgrading the version of the `sdf-cli`
which now outputs cached dependency in std-out in their dependency
order.

It also enables asset selection, via sdf's `--targets-only` flag and
passing in the fully qualified names of selected assets as targets to
the `sdf run` or `sdf test` commands. There are some hoops to jump
through with regards to asset checks, but this logic will be cleaned up
on the `sdf-cli` in future releases.

Bonus:
1. Duplicate Descriptions (this is fixed with the latest release of the
`sdf-cli`)
2. The `--log-form` flag is not necessary with the latest release

## How I Tested These Changes

Tests for caching have been added. Asset selection is covered by
existing tests.