[DA][2/n][user-code] Allow custom AutomationConditions to cross a serdes boundary (#23535)
## Summary & Motivation

If users create a custom AutomationCondition, it will of course not be serializable. This updates the serialization logic to detect this case and convert the entire AutomationCondition into a format that retains the essential information but does not contain an evaluation function. This allows us to render it in the UI (if we so desire -- currently we only render historical values, not the current condition on the definition object).

Note that this serialization logic is happening in a somewhat-surprising spot, which is the AutoMaterializePolicy object. This is because we store AutomationConditions wrapped inside an AutoMaterializePolicy on the ExternalAssetNode. I'll move this logic over to ExternalAssetNode further up the stack.

## How I Tested These Changes