[DA][6/n][user-code] Add `get_empty_slice` to AutomationContext (#23585)
## Summary & Motivation

This is commonly used throughout our built-in AutomationConditions, and likely will be commonly-used in user-defined AutomationConditions. This is significantly more ergonomic than the previous setup. Note that there is not a corresponding utility provided for getting the "full" asset slice. This is because in almost all cases where you would want to do that, you should instead just use `context.candidate_slice` to avoid returning a superset of your candidates.

Also renamed `create_empty_slice` on `AssetGraphView`, as it's inconsistent naming-wise from the other methods (in which `get_*` indicates a cheap operation).

## How I Tested These Changes