[DA][8/n][user-code] Add simplified string-only `cursor` property (#23596)
## Summary & Motivation

Technically, the Declarative Automation cursoring system can accept and store any serializable value in its "extra_state". This is useful as a performance optimization (it is common for us to want to keep track of serialized AssetSubsets between ticks, so not needing to do a double serialization / deserialization process is pretty valuable), but this overly-broad construct is likely more confusing than anything if we expose it as a public API.

Instead, expose a simplified, string-only cursor interface on the context. This mirrors the existing sensor APIs more closely.

## How I Tested These Changes