[DA][14/n][cleanup] Remove old backcompat logic (#23684)
## Summary & Motivation

Ran into this big chunk of code while working on some other changes. It is no longer necessary and can be removed.

Put simply, we had a very old format for serializing information about AMP evaluations. When we switched formats, I created some fairly extensive / hacky backcompat logic to convert from the old format to the new format automatically in the serdes layer. Over time, this has broken down to a certain extent. For example, a change several months ago made it so that this logic would no longer work properly with partitioned assets.

This is ok, as we delete old evalaution data after 30 days, and most people have updated past this version long ago (the original backcompat changes were ~9 months ago). The net impact will be that now if someone is in this (exceedingly rare) situation, their old evaluations will not have useful data on them.

Note that I am un-whitelisting a lot of classes here, which seems scary, but they were only ever serialized in the context of this class which I am now completely ignoring in the unpack step. Some existing tests actually have serialized versions of these un-whitelisted classes in them, and the whole object deserializes without error because we don't even attempt to deserialialize the innards of AutoMaterializeAssetEvaluation.

## How I Tested These Changes