make `job_name` param on `PartitionArgs` and `PartitionNames` args optional (#23983)
Having `job_name` as a required param causes problems when a pre-1.8.2
agent is used, because the agent strips the parameter off when
reserializing the object.

Added unit tests. Verified that these tests pass when cherry-picking
onto the 1.8.2 branch.

Fixes a bug that caused an error when loading the launchpad for a
partition, when using in Dagster+ with an agent with version below
1.8.2.