[DOC-353] Alerts in Dagster+ (#23963)
## Summary & Motivation

This PR creates a script to generate a guide for using alerts with a
variety of different services.

At first glance, this may seem a bit over-engineered, but the basic
issue I ran into is that there is a cross-product of different alert
types and tools which can be used together. It's really not that hard to
auto-generate all these things, so why not do it?

This means that if we ever add new notification services or alert types,
you just need to fill out a little bit of information and everything
gets populated nicely.

Anecdotally, I found the existing alerting documentation quite difficult
to navigate because it was split up into many different pages. This
consolidates it into a single one and lets someone choose the tab
associated with the notification service they're interested in, and then
view the page in the context of that specific notification service.

## How I Tested These Changes

## Changelog [New | Bug | Docs]

NOCHANGELOG