[airlift] Fix DagsterOperator task->asset lookup (#23985)
## Summary

The Dagster Operator was still using the op name to figure out what
assets to materialize. It was returning an empty list. Now, we perform
that lookup using the asset tags.

Extra motivation to get
https://linear.app/dagster-labs/issue/PLUS-1325/centralize-targeting-logic-on-asset-tags-instead-of-op-tags-naming
in so we have a standard way of mapping back and forth.

Follow up from me is more aggressive errors if we dont' find any
matching assets.

## Test Plan

Added integration test which parallels the existing switcheroo test
(that relies on the naming convention) but uses the `dag_defs`,
`task_defs` methods instead of naming convention.


## Changelog [New]

`NOCHANGELOG`