Use a consistent workspace throughout a single invocation of submit_asset_run
Summary:
Ensures that the workspace and asset graph going into a given invocation of submit_asset_run are always in sync, preventing an issue that we saw while switching over to single implicit asset jobs where the collapsing of many jobs down into a single job caused issues. To make this performant and not generate a brand new asset graph for every single run, depends on the changes earlier in the stack that caches the asset graph for a given CodeLocationEntriesSnapshot object.

Test Plan: BK, do an asset backfill that logs on asset graph creation and verify that despite being property-d on each run, the asset graph is not regenerated each run