[Run Timeline] Only commit data to cache after all data is fetched. (#24067)
## Summary & Motivation

While looking at this code again I realized there's an edge case due to
the paginated logic. In particular it's possible that the user closes
the tab before we've paginated across all runs for a specific time
period. When this happens we will have partially committed data to the
cache.

Nobody reported this issue.

## How I Tested These Changes

Added a jest test that previously failed.

## Changelog [New | Bug | Docs]
NOCHANGELOG