[DA] Refactor `eager` and `on_cron` (#23529)
## Summary & Motivation

The goal here is to make the implementations of these conditions
simpler, such that it is less daunting to copy/paste/modify. Building up
an "intermediate layer" of AutomationConditions which do common /
important tasks (e.g. being true if a parent is updated) is an important
piece of the overall puzzle. These composite conditions roughly map to
"AutoMaterializeRules" in the old world. The advantage is just that
these are composed of simpler parts, which can be inspected if desired,
and modified at will.

The most controversial name here is probably "since_last_handled". Very
open to other suggestions there.

## How I Tested These Changes

## Changelog

NOCHANGELOG