[2/n][dagster-powerbi] scaffold DagsterPowerBITranslator class (#23245)
## Summary

Builds out a very barebones translator class to build specs for PowerBI
assets.

The structure of this translator is slightly different than our existing
ones, curious for feedback. Instead of implementing many functions to
define each property of an asset, it has a single function that returns
an asset spec (which calls out to other functions for each object type).

The user also supplies a translator class rather than a translator
instance. The ontology in this model is:
- A translator class defines the logic for how to translate API data
into asset specs.
- A user can subclass a translator class to override or customize this
logic.
- An instance of a translator class represents a translation "session,"
provided with any global context information (`PowerBIWorkspaceData`, in
this case).
- An invocation of `get_asset_spec` represents translating an individual
object, and is provided with local context (the object we're
translating, `PowerBIContentData`, in this case).

One hesitation is that `Type[DagsterPowerBITranslator]` is an awkward
and potentially confusing type annotation vs passing an instance.

For PowerBI (and I'd imagine many integrations), we need global context
to resolve local translation because each local object has relationships
with the global context - in this case we need to figure out the asset
key of an upstream asset from its internal PowerBI ID.

## Test Plan

n/a, subsequent PRs will test impl - this PR is mostly around settling
on the translator API