deploying to kubernetes guide (#23978)
## Summary & Motivation
Guide for how to deploy dagster OSS to a kubernetes cluster. The guide
is based on the existing guide
[here](https://docs.dagster.io/deployment/guides/kubernetes/deploying-with-helm)
and my own notes i took when deploying to a k8s cluster a while ago. I
tried to strip out anything from the existing guide that wasn't strictly
necessary, but that leaves some content that might need to be put in
other guides like:
- adding secrets (used to connect to S3 in the existing guide)
- setting up your jobs to have a kubernetes job per step 

I made a new example project to deploy because the existing guide is
op/jobs based and also requires things like an S3 account which felt
like unnecessary setup for this guide. it's a super simple project, but
i figured that the content of the project isn't the main point

Leaving my other thoughts/questions etc in PR comments

## How I Tested These Changes

## Changelog [New | Bug | Docs]

`NOCHANGELOG`