[dagster-airlift] fix cacheableasset bugs (#24147)
We attempt to retrieve the task name off of the node name, and end up
with a mismatch because of the added unique identifier. Instead, use a
generic name that doesn't fire the convention.

Previously, assetspecs which were not mapped to tasks but passed into
the builder weren't getting coerced, and this would cause errors at repo
load time.

Add testing for the latter problem, and change existing tests for the
former.