increase default value of DAGSTER_DBT_TERMINATION_TIMEOUT_SECONDS and make it configurable
Summary:
This is currently set to only 2 seconds which is very tight - you could imagine cleanly terminating a running dbt process that is doing a bunch of stuff taking a lot longer. Increase the default and make it configurable via env var.

Test Plan: Terminate a running dbt job, watch it still finish