increase default value of DAGSTER_DBT_TERMINATION_TIMEOUT_SECONDS and make it configurable (#24183)
Summary:
This is currently set to only 2 seconds which is very tight - you could
imagine cleanly terminating a running dbt process that is doing a bunch
of stuff taking a lot longer. Increase the default and make it
configurable via env var.

Test Plan: Terminate a running dbt job, watch it still finish

## Summary & Motivation

## How I Tested These Changes

## Changelog [New]

[dagster-dbt] Increased the default timeout when terminating a run that
is running a `dbt` subprocess to wait 25 seconds for the subprocess to
cleanly terminate. Previously, it would only wait 2 seconds.
[dagster-sdf] Increased the default timeout when terminating a run that
is running an `sdf` subprocess to wait 25 seconds for the subprocess to
cleanly terminate. Previously, it would only wait 2 seconds.