Fix time range partition selection input resetting when partition health data updates (#24207)
## Summary & Motivation
As titled fixes a bug where the selection input keeps resetting. This is
due to partition health data being fetched serially for many assets. As
each asset comes in that constrains the available partitions to select
from which changes the available values for the input. The problem is
that, in the case were looking at, every asset has the same partition
definition, so we end up creating a new values array which triggers the
`useEffect` that resets the text input, however the values array is a
new reference but otherwise it has the exact same values. To get around
this issue I'm json.stringifying the values array same as we already do
to the partition names array (for the same reason I suspect).

## How I Tested These Changes

Used app-proxy to make sure the repro from
https://linear.app/dagster-labs/issue/FE-536/partition-date-input-selection-resets-when-modifying-dates
goes away.

## Changelog [New | Bug | Docs]
NOCHANGELOG