avoid gRPCs when resolving partition tags for implicit asset jobs (#23885)
## Summary & Motivation

We discovered that backcompatibility for the single-implicit-asset-job
change is actually a three-body problem, between the host processes,
agent, and user code processes:
- It's now possible for a job to target asset definitions with different
`PartitionsDefinitions`
- You need a `PartitionsDefinition` to resolve the run tags for a given
partition
- This means that, to get the run tags corresponding to a particular
partition, you need to know what assets the run is targeting
- Jobs can be assigned custom functions that determine run tags for a
partition
- These custom functions mean that, at least sometimes, there needs to
be a gRPC to the user code process to resolve the tags for a partitioned
run
- Currently, we always gRPC to the user code process to resolve the tags
for a partitioned run
- In a recent change, we added an `selected_asset_keys` field to the
gRPC param to resolve the tags. We also added a required `job_name`
field, with the goal of one day far into the future being able to remove
the `partition_set_name` field.
- Then we encountered the three-body problem: the agent re-serializes
the parameter object, and older agent versions don't know about the
`selected_asset_key` or `job_name` parameters, so they gets stripped,
which means that the user code process doesn't receive it, which means
that the user code process will fail to deserialize the parameters
object (because of the required `job_name` parameter) and, even if it
could deserialize it, wouldn't know what asset keys were selected to
resolve the tags for the partition.

Ok, so what do we do? Luckily, there are only two kinds of jobs out
there:
- Jobs that definitely do not have custom tags-for-partition functions.
Currently, all jobs that might have assets with different
`PartitionsDefinition`s are implicit asset jobs, which fall into this
category.
- Jobs that might have custom tags-for-partition function.

This means that, in the case that `selected_asset_keys` are needed to
resolve partition tags, a gRPC is not needed.

So this PR does a couple things:
- It computes the partition tags on _on the host process_ for implicit
asset jobs, which means we can avoid gRPC entirely.
- It removes `selected_asset_keys` from `PartitionArgs` and
`PartitionNamesArgs`. This reflects the fact that we can't rely on it
making it across the gRPC boundary, because it get stripped out by older
agents. It's no longer needed, because the only scenarios where it used
to be needed are now scenarios where we avoid gRPC in the first place.

## How I Tested These Changes

## Changelog [New | Bug | Docs]

NOCHANGELOG