[bugfix] Fix asset selection and asset check selection handling on sensor RunRequests (#24182)
## Summary & Motivation

When a sensor targets an `asset_selection` (as opposed to a job, which
is always targeted when using any other targeting param, including
`target`), any `RunRequest` produced by its eval function undergoes an
extra step of processing, where it is turned into a `RunRequest`
targeting a base asset job.

This extra processing sets `asset_selection` on the new `RunRequest` if
it is not already set. Prior to this PR, the check for it already being
set was ignoring an `[]` value (since this is false in booleans). This
was causing the `asset_selection` on the `RunRequest` to be reset to the
`asset_selection` for the entire sensor, which is incorrect. Also,
checking for unexpected asset check keys (keys not part of the match the
sensor asset selection subset) was not being done.

This PR:

- Changes the boolean to check for `asset_selection=None` explicitly.
- Adds a check for unexpected asset check keys. This emits a warning
instead of an error if unexpected keys are detected to avoid breakage.
- Changes the `asset_selection` for the dagster-airlift sensor to target
all asset checks instead of the empty set of assets. The empty set of
assets worked incidentally, targeting all asset checks actually makes
sense. Note that this will continue to work against older versions of
Dagster.

Fixes
https://linear.app/dagster-labs/issue/FOU-376/make-airflow-polling-sensor-work-with-asset-selection=.

## How I Tested These Changes

Added unit tests.

Made sure this test in dagster-airlift started passing with the
`asset_selection=*` change:

```
tutorial_example_tests/integration_tests/test_peer_with_check_e2e.py::test_peer_reflects_dag_completion_status_and_runs_check
```

## Changelog [Bug]

Fixed a bug where setting `asset_selection=[]` on `RunRequest` objects
yielded from sensors using `asset_selection` did not work correctly.