Refresh "What's Dagster+" (#24007)
Aug 30th Lopp: 

This PR has been updated to incorporate a few changes (sorry for the
large size 😢 ):

- The Dagster+ getting started topology now has two docs: "what's
dagster+" and "getting started". This removes the dagster+ features doc.
- Getting started focuses on being an actionable guide, without trying
to re-describe a lot of what is in the product NUX. It relies fairly
heavily on the adding a code location guide instead of replicating the
content
- What's Dagster+ builds off of Jordan's original content to explain
briefly what Dagster+ is to a audience familiar with Dagster OSS, and
also adds the short list of Dagster+ features (hence the removal of the
features doc). This was also a natural place to put cross link out to
parts of the dagster.io site as well as other guides
- As a result, I decided to flatten the overall structure of the
dagster-plus docs folder. "What's dagster+" and "getting started" are
placed at the top level. Code requirements is moved to the deployment
section, allowing us to remove the entire getting-started directory.
- This PR also removes the "deployment types" doc in favor of the brief
description of Serverless vs Hybrid in the getting started / whats
dagster+ docs. Fixes the links as a result of this removal.
- Adds the hybrid doc landing page, which takes the rest of Jordan's
original content

---------

Co-authored-by: Sean Lopp <lopp@elementl.com>