[docs-beta] Make all MDX files render in Docusaurus (#24272)
## Summary & Motivation

This updates the MdxWriter Sphinx builder to be able to handle all Dagster API documentation.
Some parts are still not implemented, but the generated MDX is valid and can be rendered by Docusaurus.
Next up, translating the missing pieces.

I removed old deprecated Airbyte API docs as part of this change. These are unmaintained APIs that are no longer being updated and should not be documented.

## How I Tested These Changes

In `sphinx`:

`make mdx`
`make copy_mdx`

In `docs-beta`:

`yarn start`

## Changelog

NOCHANGELOG