Change logger behavior so that setting a default logger in a Definitions object actually sticks, without needing to set run config
Summary:
Right now if you customize a logger on a job, you also need to set additional configuration on the run when launching the job in order for it to actually use that logger and not fall back to the detault system logger. This is quite different than what happens when you set an executor on a job, and is generally confusing.

This PR instead makes setting logger config more like executor config - if you set it, it is assumed that you want to use that logger, and applies that configuraiton as the default even if no configuration is set.

The main blocker here is probably that this could be considered a breaking change - before, when you specified a logger on a job you were giving your self the option of opting into it, and now you are overriding whatever loggers you set as the default loggers.