only retry asset run submission on DagsterInvalidSubsetError raised within a code server process, not all code server errors
Summary: https://github.com/dagster-io/dagster/pull/21631 was a little overzealous and assumed that any error raised during execution plan creation was retryable, but there are several that are not (for example, missing config, missing env vars, etc.). This PR walks that back and only triggers the retry on the invalid subset errors specifically.

Test Plan: We had a lovely automated test for this that was sadly reverted in https://github.com/dagster-io/dagster/pull/23777. For now, manually added raising a DagsterInvalidSubsetError and another error in the ExecutionPlanSnapshot grpc server call - verified that the former is retried when running an asset backfill and the latter is not.