Extend adls2 pickle io manager config (#24175)
This PR adds the 'lease_duration' attribute to the ADLS2PickleIOManager 

## Summary & Motivation
This PR is a followup of #21150. The attribute 'lease_duration' is added
to ADLS2PickleIOManager, so that it can be configured when the IO
Manager is used as a dagster resource.

## How I Tested These Changes

Successfully run dagster-azure test with 20 passed and 3 skipped.