refactor step context partition methods (#24293)
## Summary & Motivation

This is a non-behavior-changing PR that refactors some methods within
the internal execution context classes. This aim of this PR is to
prepare for https://github.com/dagster-io/dagster/pull/23824, by
centralizing the logic for getting the `PartitionsDefinition` for an
execution.

Working on this made me aware of a behavioral weirdness that I think we
should consider addressing in 1.9: `context.partition_keys` etc. will
return a non-`None` value in steps that target no partitioned assets, if
other steps in the same run do target partitioned assets.

## How I Tested These Changes

## Changelog

NOCHANGELOG