[dagster-fivetran] Pull column schema at runtime after syncing table (#23631)
## Summary


Currently, we attach column schema information to Fivetran asset
definitions [if provided by the API,
but](https://fivetran.com/docs/rest-api/connectors#retrieveconnectordetails)

> The schema config includes schemas, tables, and columns. If you have
not specified any schema configurations that differ from the default,
the response will consist of only the top schema level.

In other words, if a user is syncing an entire table, we don't get this
information at Asset definition creation time. This is why our Fivetran
assets on Hooli don't have schema info.

We can separately request this information from the API, but it requires
one request per table, which adds up quickly. This PR
adds a default-on option to fetch column schema information after
syncing a set of Fivetran tables, so we only fetch the info when a table
is newly materialized.

In the future, when we rework this API's structure, we should consider
moving this to a deferred metadata fetch method like we do on dbt.

## Test Plan

New unit tests. Tested against live Fivetran.