[ui] Disable sensor cursor editing based on code location permissions (#24436)
## Summary & Motivation

Disable the cursor editing button on sensors for users who don't have
permission.

Also fixed the tooltip width on the "Reset sensor status" button,
because with a very long object name, the tooltip is comically wide.

## How I Tested These Changes

View a sensor with viewer permissions, verify that the buttons are
disabled and tooltip'd.

## Changelog

- [ ] `BUGFIX` [ui] Disabled sensor cursor edit button for
non-permissioned users. Edits would fail on the backend anyway, but the
disabled button is clearer.