Correctly account for end_offset when deciding which partition key to look for in a freshness-based asset check
Summary:
The current logic returns the first partition window that ends before the deadline, but I think what we actually want here is the most recent partition at the time of the deadline. The former does not correctly account for the end_offset shifting over the last partition window - the latter does. This PR makes that change.

Resolves https://github.com/dagster-io/dagster/issues/24394.

Test Plan: Adding a new test case now