[ui] Fix some bugs in the code location UI (#24480)
## Summary & Motivation

- Make the code location status column a little wider on the code
location list view.
- Improve the empty state for location entries that are in an error
state. If there are errors, show a button to view them directly from the
empty state, rather than suggesting the user go hunt for errors.

## How I Tested These Changes

View code location list and code location page with a Python error in
place. Verify the above.

## Changelog

- [ ] `NEW` [ui] When landing on an empty state for an object in a code
location with errors, show a button to view the error directly on that
page.