[DA] Add `on_missing` prebuilt condition (#24225)
## Summary & Motivation

Adds a new static constructor for a fairly-common situation. Users want new partitions to get filled in as soon as possible, but never want a partition to get re-materialized.

You could achieve that behavior with AutomationCondition.eager() & AutomationCondition.missing(), but that would result in a much more complex policy than necessary. This is a much cleaner (and more performant) option, and gives first-class support to a common use case

## How I Tested These Changes

## Changelog [New]

Added `AutomationCondition.on_missing()`, which materializes an asset partition as soon as all of its parent partitions are filled in.