[ui] Permission check for reporting materialization event (#24571)
## Summary & Motivation

Use the `hasReportRunlessAssetEventPermission` field added in https://github.com/dagster-io/dagster/pull/24564 to set the disabled state of controls that allow reporting the materialization.

This includes disabling the menu item in the action assets menu, instead of just disabling the dialog button.

## How I Tested These Changes

Force the permission to be false, verify that the menu item is disabled with a tooltip.

## Changelog

- [ ] `BUGFIX` Use the correct permission to disable the "Report materialization event" button for assets.