[3/n][dagster-sigma] Build out organization parsing logic (#24477)
## Summary

Introduces  `build_organization_data`, which does the under-the-hood heavy lifting of building the Sigma object model.

There are two main object types we care about in Sigma's ontology:  [datasets](https://help.sigmacomputing.com/docs/datasets).
 and [workbooks](https://help.sigmacomputing.com/docs/workbooks-overview).

Datasets are the Sigma equivalent of Power BI Semantic Models, they are views (optionally materializable, which we might want to handle later) which sit on top of tables or other data sources.

Workbooks consist of several pages composed of elements, each of which is a table or other visualization built from a dataset.

Sigma's APIs are a little cumbersome here - we'd like to build a relationship of dependencies from workbooks to datasets and datasets to raw source data, but few APIs exist for datasets. It sounds like these will be available in the future once the migration to [data models](https://help.sigmacomputing.com/docs/intro-to-data-models) is complete, but these don't have APIs yet.

For now, our procedure is:

- Build linkages from workbooks to the datasets they depend on
- Inspect the data columns used in workbooks to assemble a partial view of datasets' column schema (which is not directly available via API)
- Inspect the SQL queries used in each "element" in a workbook using `sqlglot`. Each source table referred to by the SQL query is a source table used by the dataset linked to that element.

## Test Plan

Unit test validating that our model of organization data matches expectations. Plan to add more complex unit tests.


## Changelog

> NOCHANGELOG