[cleanup] Quarantine ValidAssetSubset class (#23843)
## Summary & Motivation

The ValidAssetSubset class existed to enable us to do operations over subsets of assets that we knew to be valid. The AssetSlice class does the same thing, making ValidAssetSubset redundant and a lot of code to comprehend.

Removing this entirely would require updating all the internal auto_materialize_rule_impls, so for now this just hides the ValidAssetSubset in a legacy folder and reduces dependence on it throughout the rest of the codebase.

## How I Tested These Changes

## Changelog [New | Bug | Docs]

NOCHANGELOG