AssetSlice -> EntitySlice (#23910)
## Summary & Motivation

Similar to the previous PR, but this time for AssetSlice. This lets us create and use slices of arbitrary EntityKeys. To help make this a bit simpler, I refactored the internal representation of an AssetSlice / EntitySlice to store a separate key/value so that the type system could independently understand the key type. I also removed some unused methods / code paths to aovid needing to refactor them . In particular, I removed the `time_windows` property of AssetSlice as it was unused outside of tests. We can always bring that back if we find a use for it.


## How I Tested These Changes

## Changelog [New | Bug | Docs]

NOCHANGELOG