Generic AutomationContext, AutomationResult, AutomationCondition (#24111)
## Summary & Motivation

This updates all of the AutomationCondition-adjacent objects to be generic, allowing them to operate over any type of EntityKey (if applicable). This opens the door to creating AutomationConditions that operate over AssetCheckKeys.

Note that I had to divest from `@record` due to an issue with how they handle generics: https://github.com/dagster-io/dagster/pull/23832

In its place, I went with subclassing our old friend `DagsterModel`. The reasoning here is that pydantic handles generics / hashing / inheriting fields in a better way than `@dataclass`. The main concern with this in the past was perf, but we're not actually constructing many instances of the conditions themselves, so that's not a huge concern here.

I also created a separate `BuiltinAutomationCondition` class for our built-in conditions to avoid "polluting" the `AutomationCondition` class that users will subclass from (pydantic basemodels put a lot of constraints on objects that would be pretty unintuitive to run into).

## How I Tested These Changes

## Changelog [New | Bug | Docs]

NOCHANGELOG