[rename] EntitySubset -> SerializableEntitySubset, EntitySlice -> EntitySubset (#24200)
## Summary & Motivation

As title, this renames EntitySubset to SerializableEntitySubset, which frees up the EntitySubset name for EntitySlice.

In general, SerializableEntitySubset is best treated as a pure serialization format, which should swiftly be converted back into an object enriched with an AssetGraphView as soon as possible. If that's the case, we should use the "EntitySubset" name for the thing that users will actually be interacting with, as "Slice" is not as accurate of a name.

## How I Tested These Changes

## Changelog [New | Bug | Docs]

NOCHANGELOG