[refactor] Remove CachingStaleStatusResolver from AssetGraphView (#24241)
## Summary & Motivation

As title. This simplifies the process of constructing an AssetGraphView -- you just need to pass in an AssetGraph and a DagsterInstance.

Currently, no stale status calculation is done within the view object, but if/when we change that, we'll construct the object responsible for doing that querying within the AssetGraphView.

## How I Tested These Changes

## Changelog [New | Bug | Docs]

NOCHANGELOG