Add sync capabilities to DataLoader class (#24323)
## Summary & Motivation

This PR adds sync support for the DataLoader pattern.

The pattern itself is a very clean way of allowing us to avoid creating separate BatchXLoader objects and threading them through a bunch of different functions. However, it's not realistic (at least in the short term) to have all callsites which require this sort of caching behavior to be async. So adding this capability allows sync code to take advantage of a shared cache without major refactors.

One explicit choice made here is that it is *not* required to call "prepare" before doing a synchronous query, which differs a bit from our existing BatchXLoader objects, which require you to add your keys beforehand, otherwise they'll error.

Personally, I think this works out a lot nicer in practice, as it avoids an entire class of runtime errors, at the cost of making it a bit less visible if you do something inefficient. In reality, a lot existing callsites just add the key right before doing the query, which just feels painful.

Needed to add a separate _blocking_batch_load method because it's not possible to turn an async function into a sync function in a completely safe way.

## How I Tested These Changes

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_